splice the common libc functions and ctypes throughout the library
