+++ src/unix/hermit/mod.rs
@@ -696,6 +696,12 @@ f! {
 }
 
 extern {
+    pub fn abs(i: ::c_int) -> ::c_int;
+    pub fn atof(s: *const ::c_char) -> ::c_double;
+    pub fn labs(i: ::c_long) -> ::c_long;
+    pub fn rand() -> ::c_int;
+    pub fn srand(seed: ::c_uint);
+
     pub fn bind(s: ::c_int, name: *const ::sockaddr, namelen: ::socklen_t)
         -> ::c_int;
