+++ src/unix/notbsd/emscripten.rs
@@ -1573,6 +1573,12 @@ f! {
 }
 
 extern {
+    pub fn abs(i: ::c_int) -> ::c_int;
+    pub fn atof(s: *const ::c_char) -> ::c_double;
+    pub fn labs(i: ::c_long) -> ::c_long;
+    pub fn rand() -> ::c_int;
+    pub fn srand(seed: ::c_uint);
+
     pub fn setpwent();
     pub fn endpwent();
     pub fn getpwent() -> *mut passwd;
